Former-commit-id: 29677326c7a0b9dcf115b06d8cb2e0645a2c7614 [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee] [formerly 29677326c7a0b9dcf115b06d8cb2e0645a2c7614 [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee] [formerly 6a1714a7b00a5510b395a91ab3d4e4170651907a [formerly f65e8a7253c65b5d59e78ed57e3d73bf01b2e06f]]]
Former-commit-id: 6a1714a7b00a5510b395a91ab3d4e4170651907a
Former-commit-id: e280c293be772e3b76431146949849c8b59106f3 [formerly 44ecd4408b0a495b887d8514b0d31c9c7be47f74]
Former-commit-id: 6af9e88e9374794fdb017013e2bcc49bbaf728f2

+++ edexOsgi/com.raytheon.uf.common.dataplugin.acarssounding/src/com/raytheon/uf/common/dataplugin/acarssounding/tools/Airport.java
@@ -32,6 +32,7 @@ import javax.xml.bind.annotation.XmlElement;
  * Date         Ticket#    Engineer    Description
  * ------------ ---------- ----------- --------------------------
  * Apr 16, 2009            jkorman     Initial creation
+ * Jul 23, 2014 3410       bclement    location changed to floats
  * 
  * </pre>
  * 
@@ -46,10 +47,10 @@ public class Airport {
     private String id;
 
     @XmlElement
-    private Double latitude;
+    private Float latitude;
 
     @XmlElement
-    private Double longitude;
+    private Float longitude;
 
     @XmlElement
     private Double elevation;
@@ -98,7 +99,7 @@ public class Airport {
     /**
      * @return the latitude
      */
-    public Double getLatitude() {
+    public Float getLatitude() {
         return latitude;
     }
 
@@ -106,14 +107,14 @@ public class Airport {
      * @param latitude
      *            the latitude to set
      */
-    public void setLatitude(Double latitude) {
+    public void setLatitude(Float latitude) {
         this.latitude = latitude;
     }
 
     /**
      * @return the longitude
      */
-    public Double getLongitude() {
+    public Float getLongitude() {
         return longitude;
     }
 
@@ -121,7 +122,7 @@ public class Airport {
      * @param longitude
      *            the longitude to set
      */
-    public void setLongitude(Double longitude) {
+    public void setLongitude(Float longitude) {
         this.longitude = longitude;
     }
